+++ tests/unit/actions/ActionsMapTest.php
@@ -35,29 +35,6 @@ class ActionsMapTest extends BaseTestCase
         $this->assertSame(42, $baz->made());
     }
 
-    public function testRunForNode()
-    {
-        $map = new ActionsMap([
-            'foo' => function (Token $foo) {
-                return $foo->getContent();
-            },
-            'Bar' => function (TreeNodeInterface $bar, TreeNodeInterface $foo) {
-                return '[' . $foo->made() . ']';
-            },
-            'Baz' => ActionsMap::DO_BUBBLE_THE_ONLY,
-        ]);
-
-        $foo = new Token('foo', 'lorem ipsum');
-        $foo->make($map->runForNode($foo));
-        $bar = new NonTerminal('Bar', [$foo]);
-
-        $this->assertEquals('[lorem ipsum]', $map->runForNode($bar));
-        $this->assertNull($map->runForNode(new Token('x', '42')));
-
-        $baz = new NonTerminal('Baz', [$foo]);
-        $this->assertEquals('lorem ipsum', $map->runForNode($baz));
-    }
-
     public function testThrowingInTerminal()
     {
         $map = new ActionsMap([